refactor: centralize topic constants
