+++ handlers/blogs/types.go
@@ -1,0 +1,12 @@
+package blogs
+
+import (
+	corecommon "github.com/arran4/goa4web/core/common"
+	hcommon "github.com/arran4/goa4web/handlers/common"
+)
+
+// IndexItem exposes the navigation item type.
+type IndexItem = corecommon.IndexItem
+
+// CoreData exposes the common.CoreData type for handlers.
+type CoreData = hcommon.CoreData
