--- handlers/search/consts.go
@@ -1,13 +1,0 @@
-package search
-
-// BloggerTopicName is the default name for the hidden blogger forum.
-const BloggerTopicName = "A BLOGGER TOPIC"
-
-// LinkerTopicName is the default name for the hidden linker forum.
-const LinkerTopicName = "A LINKER TOPIC"
-
-// WritingTopicName is the default name for the hidden writing forum.
-const WritingTopicName = "A WRITING TOPIC"
-
-// Alphabet lists letters for search navigation.
-const Alphabet = "abcdefghijklmnopqrstuvwxyz"
