+++ startup_test.go
@@ -6,6 +6,7 @@ import (
 	"testing"
 
 	"github.com/DATA-DOG/go-sqlmock"
+	hcommon "github.com/arran4/goa4web/handlers/common"
 )
 
 func TestEnsureSchemaVersionMatch(t *testing.T) {
@@ -20,7 +21,7 @@ func TestEnsureSchemaVersionMatch(t *testing.T) {
 	mock.ExpectQuery(regexp.QuoteMeta("SELECT COUNT(*) FROM schema_version")).
 		WillReturnRows(sqlmock.NewRows([]string{"cnt"}).AddRow(1))
 	mock.ExpectQuery(regexp.QuoteMeta("SELECT version FROM schema_version")).
-		WillReturnRows(sqlmock.NewRows([]string{"version"}).AddRow(ExpectedSchemaVersion))
+		WillReturnRows(sqlmock.NewRows([]string{"version"}).AddRow(hcommon.ExpectedSchemaVersion))
 
 	if err := ensureSchema(context.Background(), db); err != nil {
 		t.Fatalf("ensureSchema: %v", err)
@@ -42,7 +43,7 @@ func TestEnsureSchemaVersionMismatch(t *testing.T) {
 	mock.ExpectQuery(regexp.QuoteMeta("SELECT COUNT(*) FROM schema_version")).
 		WillReturnRows(sqlmock.NewRows([]string{"cnt"}).AddRow(1))
 	mock.ExpectQuery(regexp.QuoteMeta("SELECT version FROM schema_version")).
-		WillReturnRows(sqlmock.NewRows([]string{"version"}).AddRow(ExpectedSchemaVersion - 1))
+		WillReturnRows(sqlmock.NewRows([]string{"version"}).AddRow(hcommon.ExpectedSchemaVersion - 1))
 
 	if err := ensureSchema(context.Background(), db); err == nil {
 		t.Fatalf("expected error")
